Errors in delete account, admin check

+++ my_app/views.py
@@ -17,14 +17,20 @@ from django.core.mail import send_mail
 from django.conf import settings
 from datetime import date
 from datetime import timedelta, datetime
+from django.http import HttpResponseRedirect
 
 # def logout_view(request):
 #     logout(request)
 #     return redirect('registration/login.html')
+
 @login_required
 def delete_account(request):
     #username = request.user.username
+    confirm_name = request.GET.get('name', '')
     user = request.user
+    if (Group.objects.filter(admin_user=user).exists()):
+        messages.error(request, confirm_name + "Please first transfer ownership of your group or delete the group before deleting your account.")
+        return HttpResponseRedirect(request.META.get('HTTP_REFERER', '/'))
     user.delete()
     return redirect('home')
 #
@@ -305,9 +311,10 @@ def create_group(request):
 
     if UserJoinGroup.objects.filter(user=user).exists():
         # any popup?
+        print("eete")
         return render(request, 'group_settings.html')
     
-    group = Group(name=group_name)
+    group = Group(name=group_name, admin_user=user)
     group.save()
     # Check if there's an existing relationship between user and category
     print(group.name)
